[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -2,11 +2,11 @@
 
 int main(void) {
 
-  int A, B;
+  long long int A, B;
 
-  scanf("%d %d", &A, &B);
+  scanf("%lld %lld", &A, &B);
 
-  printf("%d", (A+B)*(A-B));
+  printf("%lld", (A+B)*(A-B));
   
   return 0;
 }
